requirements.txt
python-telegram-bot==12.7
emoji==1.7.0
psycopg2==2.9.3